pyproject.toml
[tool.poetry]
name = "chapter_tutorial"
version = "0.1.0"
description = ""
authors = ["juhwanlee-diquest <juhwanlee@diquest.com>"]

[tool.poetry.dependencies]
python = "^3.8"
apache-airflow = "^2.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
